Struct loxide::compiler::Compiler

pub struct Compiler<'a> {
    lexer: Lexer<'a>,
    chunk: Chunk,
    had_error: bool,
    in_panic_mode: bool,
    peeked_token: Option<Option<Token>>,
}

The data type which manages compilation from source code to bytecode.

Fields
lexer: Lexer<'a>chunk: Chunkhad_error: boolin_panic_mode: boolpeeked_token: Option<Option<Token>>
Implementations

impl<'a> Compiler<'a>

pub fn new(lexer: Lexer<'a>) -> Self

pub fn compile(self) -> Result<Chunk, ()>

Compile the source code represented by the given Lexer into bytecode, using a single-pass Pratt parsing technique.

fn finalize_and_return(self) -> Result<Chunk, ()>

After compilation has finished, run any remaining code and determine whether to return with success or an error.

fn parse_with_precedence(&mut self, precedence: InfixOperatorPrecedence)

Parse the upcoming tokens using a given infix operator precedence. We expect the next token to be a prefix, such as a number or a minus.

After parsing the prefix expression, we try to parse infix expressions of the given precedence or higher, returning once the precedence is too low.

fn compile_expression(&mut self)

Compile top level expressions.

fn compile_grouping(&mut self, token: Token)

Compile parenthesis groupings.

fn compile_number(&mut self, token: Token)

fn compile_unary_expression(&mut self, token: Token)

Compile expressions that start with unary operators like minus.

fn compile_binary_expression(&mut self, token: Token)

Compile expressions that contain binary infix operators like plus or star. The passed in token is expected to be the infix operator. Everything before the infix operator has already been parsed.

fn add_constant(&mut self, value: Value, token: Token) -> u8

fn next_token(&mut self) -> Option<Token>

fn peek_next_token(&mut self) -> Option<&Token>

fn expect_token_type(
&mut self,
expected_token_type: TokenType,
message: &str
) -> Option<Token>

Report an error if the next token isn’t of the given token type. Return the token if it is of the given type. Otherwise don’t consume it and return None.

fn create_eof_token(&self) -> Token

fn report_error<'t>(&mut self, token: impl Borrow<Token> + 't, message: &str)

Report an error with the given message on the given token. If we’re in panic mode, suppress the error. Otherwise, report the error, enable panic mode, and set an internal flag saying we had an error.

fn line_number(&self, token: &Token) -> usize

fn token_string(&self, token: &Token) -> &str

fn get_prefix_compilation_function(
&self,
token_type: &TokenType
) -> Option<fn(_: &mut Compiler<'a>, _: Token)>

fn get_infix_compilation_function(
&self,
token_type: &TokenType
) -> Option<fn(_: &mut Compiler<'a>, _: Token)>

fn print_error(&mut self, format_args: Arguments<'_>)